Split default merge and createCustomMerge
